Avoid recursion stack overflow with nested refs
Fixes [cpan #97526]: stack overflow with moderately complex data structures
and adds a max ref limit of 4000.

Unfortunately we cannot precisely count the stack depth of nested
array->store->ref->array->store->ref calls, without a costly hash.
pseen/hseen is bypassed here, those refs are all fresh.
So we just count the number of refs which can recurse, and error
when we know that we recurse (entry==1) and the number of refs exceeds
a certain threshold (4000). This is dependent on the platform, ulimit -s
and threaded.

Note we cannot store that many nested refs anymore, which we could store
before. But we need to escape the stack overflow.
Before we could do
  perl -e'use Storable qw(dclone); $t = [$t] for 1..21830; dclone $t'
now we will fail at
  perl -e'use Storable qw(dclone); $t = [$t] for 1..4002; dclone $t'
Max. recursion depth of nested refs exceeded at -e line 1.