improve parallel build races
the INST_DYNAMIC target removes temp. the extension, which will lead to
failure of loading the current Storable.so, rather it loads an older
one.
we could remove the $(RM_F) $@ line in dynamic_lib, but waiting is good
enough. the first probes are not that important, they will stabilize later.

fix building in core with miniperl, which cannot load the extension. so
defer.

probe for a usable Storable extension beforehand, not to emit a wrong
bisected stacksize.h. this is only a problem in CORE with miniperl.